requirements.txt
#Requirements
tensorflow
flask
pillow
numpy
pandas
tensorflow-hub
tensorflow_hub